Fix compiler warnings throughout the app.

+++ buyvmmanager/BVMIPListViewController.m
@@ -33,7 +33,7 @@
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
-    return self.ips.count;
+    return (NSInteger) self.ips.count;
 }
 
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath
